+++ callingcardstools/PeakCalling/yeast/poisson_pval_vectorized.py
@@ -1,17 +1,19 @@
 import logging
 
-from numpy import int64
 from pandas import Series
 from scipy.stats import poisson
 
 logger = logging.getLogger(__name__)
 
 
-def poisson_pval_vectorized(total_background_hops: Series,
-                            total_experiment_hops: Series,
-                            background_hops: Series,
-                            experiment_hops: Series,
-                            pseudocount: float = 1e-10) -> Series:
+def poisson_pval_vectorized(
+    total_background_hops: Series,
+    total_experiment_hops: Series,
+    background_hops: Series,
+    experiment_hops: Series,
+    pseudocount: float = 0.1,
+    **kwargs
+) -> Series:
     """
     Compute the Poisson p-value for the given hops counts.
 
@@ -27,7 +29,7 @@ def poisson_pval_vectorized(total_background_hops: Series,
     :param experiment_hops: a pandas Series (column of a dataframe)
         of number of hops in the experiment by promoter region.
     :type experiment_hops: Series[int64]
-    :param pseudocount: , defaults to 1e-10
+    :param pseudocount: , defaults to 1.0
     :type pseudocount: float, optional
     :return: a pandas Series of length equal to the input Series with the
         Poisson p-value for each row.
@@ -55,23 +57,35 @@ def poisson_pval_vectorized(total_background_hops: Series,
     array([0.01438768, 0.00365985, 0.00092599])
     """
     # check input
-    if not len(total_background_hops) == len(total_experiment_hops) == \
-            len(background_hops) == len(experiment_hops):
-        raise ValueError('All input Series must be the same length.')
-    if total_background_hops.min() < 0 \
-            or total_background_hops.dtype != 'int64':
-        raise ValueError(('total_background_hops must '
-                          'be a non-negative integer.'))
-    if total_experiment_hops.min() < 0 \
-            or total_background_hops.dtype != 'int64':
-        raise ValueError(('total_experiment_hops must '
-                          'be a non-negative integer'))
-    
-    # cast to `float` b/c of scipy
-    hop_ratio = (total_experiment_hops 
-                 / (total_background_hops + pseudocount)).astype('float')
-    mu = ((background_hops * hop_ratio)
-          + pseudocount).astype('float')
-    x = (experiment_hops + pseudocount).astype('float')
+    if (
+        not len(total_background_hops)
+        == len(total_experiment_hops)
+        == len(background_hops)
+        == len(experiment_hops)
+    ):
+        raise ValueError("All input Series must be the same length.")
+    if total_background_hops.min() < 0 or total_background_hops.dtype != "int64":
+        raise ValueError(("total_background_hops must " "be a non-negative integer."))
+    if total_experiment_hops.min() < 0 or total_background_hops.dtype != "int64":
+        raise ValueError(("total_experiment_hops must " "be a non-negative integer"))
 
-    return 1 - poisson.cdf(x, mu)
+    # cast to `float` b/c of scipy
+
+    # note that read_in_background_data and read_in_experiment_data in
+    # read_in_data.py require that there be at least 1 hop in both the background
+    # and the experiment. Therefore the total_background_hops and total_experiment_hops
+    # is always defined
+    hop_ratio = (total_experiment_hops / total_background_hops).astype("float")
+
+    # It is possible that there are promoters with no background hops. Add a small
+    # pseudocount to require that mu > 0, per poisson definition
+    mu = ((background_hops + pseudocount) * hop_ratio).astype("float")
+
+    # there has been a pseudocount added to experiment hops. Not necessary, removed
+    # 20240624
+    # The way this is calculated, with pyranges and a sum, this value will always be
+    # at minimum 0
+    x = experiment_hops.astype("float")
+
+    # return the p-value
+    return Series(1 - poisson.cdf(x, mu))